Moved stack module into utils and added softlink (for backwards compatibility)
